window test

+++ src/math/mod.rs
@@ -2,4 +2,4 @@ pub mod color;
 pub mod matrix;
 pub mod point;
 
-pub type Radians = f64;
+pub type Radians = f32;
